# swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080/","description":"Generated server url"}],"paths":{"/v1/sales":{"get":{"tags":["sale-controller"],"operationId":"getSales","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Sale"}}}}}}},"post":{"tags":["sale-controller"],"operationId":"createSale","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaleDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Sale"}}}}}}},"/v1/products":{"get":{"tags":["product-controller"],"operationId":"getProducts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}},"post":{"tags":["product-controller"],"operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}}},"/v1/clients":{"get":{"tags":["client-controller"],"operationId":"getUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Client"}}}}}}},"post":{"tags":["client-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Client"}}}}}}},"/v1/products/{id}":{"get":{"tags":["product-controller"],"operationId":"getClientById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/v1/clients/{id}":{"get":{"tags":["client-controller"],"operationId":"getClientById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"SaleDTO":{"type":"object","properties":{"idClient":{"type":"integer","format":"int32"},"saleDetailsDTO":{"type":"array","items":{"$ref":"#/components/schemas/SaleDetailDTO"}}}},"SaleDetailDTO":{"type":"object","properties":{"idProduct":{"type":"integer","format":"int32"},"quantity":{"type":"integer","format":"int32"}}},"Client":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"phone":{"type":"string"}}},"Product":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"description":{"type":"string"},"cost":{"type":"number","format":"double"}}},"Sale":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"client":{"$ref":"#/components/schemas/Client"},"total":{"type":"number","format":"double"},"saleDetails":{"type":"array","items":{"$ref":"#/components/schemas/SaleDetail"}}}},"SaleDetail":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"product":{"$ref":"#/components/schemas/Product"},"quantity":{"type":"integer","format":"int32"},"subtotal":{"type":"number","format":"double"}}},"ProductDTO":{"type":"object","properties":{"description":{"type":"string"},"cost":{"type":"number","format":"double"}}},"ClientDTO":{"type":"object","properties":{"name":{"type":"string"},"phone":{"type":"string"}}}}}}